Specify model to use in game name
You can specify the Amiga model to be used for each game.

If the filename contains "(A500)" or "(OCS)", an Amiga 500 with ks13 will be used.

If the filename contains "(A600)" or "(ECS)", an Amiga 600 with ks20 will be used.

If the filename contains "(A1200)" or "(AGA)", an Amiga 1200 with ks31 will be used.

if the filename does not contain any model, the configured one will be use.